Separate test commons in their own module.
